refactor: quizId 타입 long으로 변경 및 애너테이션 변경

+++ src/main/java/com/LearnDocker/LearnDocker/QuizController.java
@@ -15,7 +15,7 @@ public class QuizController {
     }
     
     @GetMapping(value="/{quizId}")
-    public String getQuizById(@PathVariable(value="quizId") int quizId) {
+    public String getQuizById(@PathVariable(value="quizId") long quizId) {
         return this.quizService.getQuizById(quizId);
     }
 }
